# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification for the AI Exposure Review backend
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /assessments:
    get:
      operationId: listAssessments
      tags: [assessments]
      summary: List assessments
      parameters:
        - in: query
          name: workspaceId
          required: false
          schema:
            type: string
            format: uuid
          description: Optional workspace filter
      responses:
        "200":
          description: Assessment list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentListResponse"
    post:
      operationId: createAssessment
      tags: [assessments]
      summary: Create an assessment
      description: Creates an assessment configuration. This scaffolds workflow metadata only and does not run a real scanner.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssessmentRequest"
      responses:
        "201":
          description: Assessment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Workspace or repository not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /assessments/{assessmentId}:
    get:
      operationId: getAssessment
      tags: [assessments]
      summary: Get assessment
      parameters:
        - $ref: "#/components/parameters/AssessmentId"
      responses:
        "200":
          description: Assessment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentResponse"
        "404":
          description: Assessment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  parameters:
...
    AssessmentId:
      in: path
      name: assessmentId
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
    ErrorDetail:
      type: object
      additionalProperties: true
    ErrorBody:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          $ref: "#/components/schemas/ErrorDetail"
      required:
        - code
        - message
    ErrorResponse:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/ErrorBody"
      required:
        - error
...
    AssessmentStatus:
      type: string
      enum: [draft, queued, running, completed, failed]
    AssessmentScannerProfile:
      type: string
      enum: [baseline]
    Assessment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        workspaceId:
          type: string
          format: uuid
        repositoryId:
          type: string
          format: uuid
        name:
          type: string
        status:
          $ref: "#/components/schemas/AssessmentStatus"
        targetBranch:
          type: string
          nullable: true
        stagingUrl:
          type: string
          format: uri
          nullable: true
        runtimeChecksEnabled:
          type: boolean
        scannerProfile:
          $ref: "#/components/schemas/AssessmentScannerProfile"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        launchedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - workspaceId
        - repositoryId
        - name
        - status
        - targetBranch
        - stagingUrl
        - runtimeChecksEnabled
        - scannerProfile
        - createdAt
        - updatedAt
        - launchedAt
    AssessmentResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Assessment"
      required:
        - data
    AssessmentListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Assessment"
      required:
        - data
    CreateAssessmentRequest:
      type: object
      properties:
        workspaceId:
          type: string
          format: uuid
        repositoryId:
          type: string
          format: uuid
        name:
          type: string
        targetBranch:
          type: string
        stagingUrl:
          type: string
          format: uri
        runtimeChecksEnabled:
          type: boolean
      required:
        - workspaceId
        - repositoryId
        - name
...